Moved all the examples from tests to examples
