+++ examples/starwars_django/tests/test_connections.py
@@ -1,9 +1,9 @@
 import pytest
 from graphql.core import graphql
 
-from .models import *
-from .schema import schema
-from .data import initialize
+from ..models import *
+from ..schema import schema
+from ..data import initialize
 
 pytestmark = pytest.mark.django_db
