+++ examples/starwars_django/tests/test_objectidentification.py
@@ -1,9 +1,9 @@
 import pytest
 from pytest import raises
 from graphql.core import graphql
-from .data import initialize
+from ..data import initialize
 
-from .schema import schema
+from ..schema import schema
 
 pytestmark = pytest.mark.django_db
